Rename gametitle to titleinfo
